backend/pyproject.toml
[project]
name = "rhr-backend"
version = "0.1.0"
description = "Raptor HR - Enterprise HR Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-allauth>=65.0,<66.0",
    "django-cors-headers>=4.4,<5.0",
    "django-filter>=24.0,<25.0",
    "drf-spectacular>=0.27,<1.0",
    "psycopg[binary]>=3.2,<4.0",
    "celery>=5.4,<6.0",
    "redis>=5.0,<6.0",
    "gunicorn>=23.0,<24.0",
    "python-decouple>=3.8,<4.0",
    "dj-database-url>=2.2,<3.0",
    "whitenoise>=6.7,<7.0",
    "requests>=2.31,<3.0",
    "cryptography>=42.0,<44.0",
    "pyjwt>=2.8,<3.0",
    "pillow>=10.0,<11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-cov>=5.0,<6.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.5,<1.0",
    "mypy>=1.10,<2.0",
    "django-stubs>=5.0,<6.0",
    "djangorestframework-stubs>=3.15,<4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"
